Disable py33 testing for now and reorder tests
There are a lot of changes needed to support py33, so it would
be a lot easier to just discontinue support for the time being.
I reorded the tests to avoid an incompatibility when pypy and
py27 in the .testrepository directory.  If pypy is run first,
the tests run successfully.

Change-Id: Ia03fb203974c7956f00c768a52ba9443c004a6c3