Validator chains now exit on the first error
Also, added some more robust error handling for buggy validators.
A broken validator will now cause Anchor to fail-closed and deny
a certificate, exactly as if that validator test had failed
normally.

A message is logged giving information about the error and a 500
is returned to the calling service with minimal information.

Change-Id: Iddbd0e831a52cc02cc569f220aebf82bd59554d0