Restore the old validator sets behaviour
This change brings back the original behaviour of validator sets,
which means each set will be tested separately. If any of the sets
passes all its validators, it will be accepted. If it passes none,
it will be rejected.

This allows multiple configurations to be served from the same
Anchor instance. For example multiple teams can configure separate
source ips allowed to apply for separate hostnames.

Closes bug: 1463748

Change-Id: Id5f811c6d46bc47ddcc1f751ef8e4e7ffd1254fd