Make sure X509_NAME lives long enough
X509_NAME_get_entry() returns a pointer into an existing name struct.
The struct must live at least as long as Entry object. That means we
have to keep a pointer to the parent for as long as Entry is alive,
otherwise it will be freed and possibly overwritten with garbage.

Change-Id: Ia5e11c0ad7202fefeca87bf59b3dedc86ccac969