Check for exception code and not type
Due to the way pecan.abort() chooses exception type, the tests cannot
rely on catching HTTPServerError for 500 errors. A number of exceptions
can be raised at random.

We have to check for code 500 and a more generic exception in this case
to succeed.

This issue doesn't apply to HTTPClientError fortunately.

Change-Id: I8231f21d50764b026c030d2fa6d0c2c092143dc1