Stop mixing IPs and domains
Domains and IPs are separate and should be treated that way. In the
alternative names, they are explicitly marked as either DNS or IP. In
the common name, IPs will be detected and checked against networks,
while domains will be checked against domains only.

Resolved domains will not be checked against the network ranges. This
is a potential security issue in case requester actually controls the
domain entries.

Change-Id: I05e49ed526331013644a78ae59d1ebcbcb8b25d5
Closes-bug: #1482058
Closes-bug: #1482059