Stop mixing IPs and domains
Domains and IPs are separate and should be treated that way. There's also
no reason to use IPs in CN anymore - it should not be checked against the
network ranges.
Resolved domains should not be checked against the network ranges either.
This is a potential security issue in case requester actually controls the
domain entries.

Change-Id: I05e49ed526331013644a78ae59d1ebcbcb8b25d5
Closes-bug: #1482058
Closes-bug: #1482059