Changed readme so that example retrieves certificate.
Made a change to the URL that the readme uses to demonstrate basic
Anchor functionality. This reflects the default CA setup in the
simple config.json that Anchor ships with today.

Note: The examples for auth/config are still outdated but at least
the README example usage works. auth/config updates can be covered
in another bug/commit.

Change-Id: I17b9580956baa5644b9a437a4da86ed232c4a76a