Created a dockerfile that runs Anchor
It works but probably doesn't run Anchor in the best way.

Once this is in the repo I can create a docker build job that
will auto generate a new upstream anchor image each time a new
merge occurs. When that is established, the process for running
the container can be simplified.

Change-Id: Ida199d4286a4b476e52d69864c97ff24633ca073