Replaced the existing Dockerfiles. Tidied up Readme
I replaced the Dockerfiles that existed, which pulled anchor
down from git with a single Dockerfile that uses the local version
of anchor.

This makes testing local changes to anchor very simple.

I also updated the README to reflect these changes and fixed the
varying line lenghts within that file.

Change-Id: I313c840f594639e6c92ed6ff067da79652b74ed0