Ignore the 'pep8.txt' file via .gitignore
Right now following the README.md and OpenStack Developer's Guide
has you go through the 'git clone' process of openstack/anchor,
and a script ('run_tests.sh') is provided which sets up your
virtual environment, installs the requisite packages, and tests.

This workflow seems to result in the creation of 'pep8.txt'
within the repository root which is (seemingly) undesirable
as it shows as an untracked, uncommitted file.

Fixed by adding the file to .gitignore but I'm fairly new to Python
so please advise if I'm somehow confused! Thanks <3

Change-Id: I7f69e40a1069fcbf17ef3f277ab1be65cf182d04