Adding bootstrap to docker container
This commit removes the key generation stage of building Anchor
containers. When a container is started it will look for a key in a
docker volume named '/key' if there is a correctly formatted key it will
attempt to use it. If not it will generate one before starting normal
operations.

Change-Id: I569f0fe07171faeed0cffc7c87f3075a094ec811